Expand exceptions

+++ app/Services/IntegrationCommonService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Enums\MarketplaceType;
 use App\Exceptions\InvalidDataException;
+use App\Exceptions\UnsuccessfulResponseException;
 use App\Models\Integration;
 use Illuminate\Support\Facades\Auth;
 
@@ -20,19 +21,28 @@ class IntegrationCommonService
         $this->marketplaceService = app('MarketplaceService');
     }
 
+    /**
+     * @throws InvalidDataException
+     * @throws UnsuccessfulResponseException
+     */
     public function createCommonIntegration(array $data)
     {
-        $marketplaceType = MarketplaceType::from($data['marketplace']);
+        $marketplaceType = MarketplaceType::tryFrom($data['marketplace']);
+        if (!$marketplaceType)
+            throw new InvalidDataException('Marketplace not found: ' . $data['marketplace']);
 
         $integrationData = [
             'username' => $data['username'],
             'password' => $data['password'],
         ];
-        $response = $this->integrationManager->createIntegration($marketplaceType->adapter(), $integrationData);
 
-        //
+        $response = $this->integrationManager->createIntegration($marketplaceType->adapter(), $integrationData);
+        if (!$response['status'])
+            throw new UnsuccessfulResponseException($response['message']);
 
         $marketplace = $this->marketplaceService->findMarketplaceByName($data['marketplace']);
+        if (!$marketplace)
+            throw new InvalidDataException('Marketplace model not found: ' . $data['marketplace']);
 
         $integrationData = array_merge($integrationData, [
             'marketplace_id' => $marketplace->id,
@@ -46,6 +56,7 @@ class IntegrationCommonService
 
     /**
      * @throws InvalidDataException
+     * @throws UnsuccessfulResponseException
      */
     public function updateCommonIntegration(Integration $integration, array $data)
     {
@@ -57,22 +68,35 @@ class IntegrationCommonService
         if (isset($data['password']))
             $integrationData['password'] = $data['password'];
 
-        if (empty($integrationData)) {
+        if (empty($integrationData))
             throw new InvalidDataException('Integration data is invalid');
-        }
 
-        $marketplaceType = MarketplaceType::from($integration->marketplace->name);
+        $marketplaceType = MarketplaceType::tryFrom($integration->marketplace->name);
+        if (!$marketplaceType)
+            throw new InvalidDataException('Marketplace not found: ' . $integration->marketplace->name);
 
-        $this->integrationManager->updateIntegration($marketplaceType->adapter(), $integration->reference, $integrationData);
+        $response = $this->integrationManager->updateIntegration($marketplaceType->adapter(), $integration->reference, $integrationData);
+        if (!$response['status'])
+            throw new UnsuccessfulResponseException($response['message']);
+
         $integration = $this->integrationService->updateIntegration($integration, $integrationData);
 
         return $integration;
     }
 
+    /**
+     * @throws InvalidDataException
+     * @throws UnsuccessfulResponseException
+     */
     public function deleteCommonIntegration(Integration $integration): void
     {
-        $marketplaceType = MarketplaceType::from($integration->marketplace->name);
-        $this->integrationManager->deleteIntegration($marketplaceType->adapter(), $integration->reference);
+        $marketplaceType = MarketplaceType::tryFrom($integration->marketplace->name);
+        if (!$marketplaceType)
+            throw new InvalidDataException('Marketplace not found: ' . $integration->marketplace->name);
+
+        $response = $this->integrationManager->deleteIntegration($marketplaceType->adapter(), $integration->reference);
+        if (!$response['status'])
+            throw new UnsuccessfulResponseException($response['message']);
 
         $this->integrationService->deleteIntegration($integration);
     }
